backend/pyproject.toml
[project]
name = "haystack-rag-app"
version = "0.1.0"
description = "RAG system with separate indexing and query services"
dependencies = [
    "fastapi>=0.115.6",
    "haystack-ai>=2.8.0",
    "markdown-it-py>=3.0.0",
    "mdit_plain>=1.0.1",
    "opensearch-haystack>=1.2.0",
    "pydantic-settings>=2.7.0",
    "pypdf>=5.1.0",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.19",
    "sentence-transformers>=3.3.1",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "mypy",
]

[build-system]
requires = ["setuptools>=75", "wheel"]
build-backend = "setuptools.build_meta"
